Make device-ios unit work with standard STF server
Alter device identity fetch to *not* go through the stf device library as it isn't up to date. This will soon go through a different mechanism with up to date IOS information.

Update the websocket screen mirroring path to be correct so that information injected via the channel name does not need to be used and video should work with a standard stf server ( albeit still connecting a modified nginx path to get to the device video ) The screen-ws-url-pattern option will still need to be set correctly from the CLI starting the device-ios unit.

Altered the paths of calls to ilibmobiledevice CLI commands to be absolute so that the path need not be set to include '/usr/local/bin' when running the device-ios unit.

Commented out the getDisplay function in device-ios/support/deviceinfo.js as well as the call to it. This is not the appropriate place to set this. It gets overwritten elsewhere so this is just a wasteful call to idevicescreenshot.