Move DeviceReady message back to device unit
The DeviceReady message being in the provider was causing a timing issue in that the provider is always running and receives a message that a device is connected from the coordinator before the stf_device_ios unit is started. As a result, a probe is sent as the result of that ReadyMessage before the device_ios unit begins listening for the probe; then the device details are never sent.

Looking back this is why the DeviceReady message was sent in original code from the device_ios unit, and it should not have been moved. This reverts back to the Android method and should solve the timing issue.

The result of the timing problem was that video width/height/url were not being set and so video did not appear.