Update model mapping
Updated the display of IOS device model information by altering the way it is looked up. Previously it was looked up by "ProductType". Now it is looked up by "Model".
Using lookup this way the storage size and color of the device can be determined.

The mapping information added was generated by scraping "The iPhone wiki" and the process is nearly automatic. Have not yet determined where to place the scripts to scrape and generate this data in the repo. They will be added as well once an appropriate place is decided.